Update navbar Search

+++ adminManageCategory.php
@@ -21,16 +21,28 @@
             // echo 123;
         }
         if(@$_GET['d']==1){
-                    $_SESSION['delete'] = true;
+                   
                     $categoryID = $_GET['categoryID'];
-                    $deleteSQL = 'DELETE FROM category WHERE categoryID = ?';
-                    $preparedeleteSQL = $GLOBALS['conn']->prepare($deleteSQL);
-                    $preparedeleteSQL->bind_param("i",$categoryID);
-                     $preparedeleteSQL->execute();
-                     $preparedeleteSQL->close();
-                    header('refresh:0; url=adminCategory.php');
+                    $checkCategorySQL = 'SELECT * FROM board WHERE categoryID = ?';
+                    $preparecheckCategorySQL = $GLOBALS['conn']->prepare($checkCategorySQL);
+                    $preparecheckCategorySQL->bind_param("i", $categoryID);
+                    $preparecheckCategorySQL->execute();
+                    $Checkresult = $preparecheckCategorySQL->get_result();
+                    $preparecheckCategorySQL->close();
+                    if($Checkresult->num_rows > 0){
+                        $_SESSION['delete'] = 2;
+                        header('refresh:0; url=adminCategory.php');
+                    }else{
+                        $_SESSION['delete'] = true;
+                        $deleteSQL = 'DELETE FROM category WHERE categoryID = ?';
+                        $preparedeleteSQL = $GLOBALS['conn']->prepare($deleteSQL);
+                        $preparedeleteSQL->bind_param("i",$categoryID);
+                        $preparedeleteSQL->execute();
+                         $preparedeleteSQL->close();
+                        header('refresh:0; url=adminCategory.php');
+                    }
+                   
         }else{
-            
         $categoryID = $_GET['categoryID'];
         $categorySQL = 'SELECT * FROM category WHERE categoryID = ? ';
         $preparecatgorySQL = $GLOBALS['conn']->prepare($categorySQL);
@@ -80,7 +92,8 @@
 </head>
 <body>
   
-    <?php require 'req/navbar.php' ?>
+    <?php require 'req/navbar.php'; if(@$_GET['d'] != 1) { ?>
+
     <div class="container-fluid  mt-3 mb-2">
     <div class="row">
         <div class="col-sm-4"> </div>
@@ -100,5 +113,6 @@
         </div>
     </div>
     </div>
+    <?php } ?>
 </body>
 </html>
